% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaemixRes_d.R
\docType{methods}
\name{[,SaemixRes_d-method}
\alias{[,SaemixRes_d-method}
\title{Get/set methods for SaemixRes object}
\usage{
\S4method{[}{SaemixRes_d}(x, i, j, drop)
}
\arguments{
\item{x}{object}

\item{i}{element to be replaced}

\item{j}{element to replace with}

\item{drop}{whether to drop unused dimensions}
}
\description{
Access slots of a SaemixRes object using the object["slot"] format
}
\keyword{methods}
